+++ LibHac/Npdm/FsPermissionBool.cs
@@ -1,4 +1,4 @@
-namespace LibHac
+namespace LibHac.Npdm
 {
     enum FsPermissionBool : ulong
     {
