+++ LibHac/Npdm/FsPermissionRw.cs
@@ -1,4 +1,4 @@
-namespace LibHac
+namespace LibHac.Npdm
 {
     enum FsPermissionRw : ulong
     {
