+++ LibHac/Npdm/KernelAccessControl.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
-namespace LibHac
+namespace LibHac.Npdm
 {
     class KernelAccessControl
     {
