+++ LibHac/Npdm/KernelAccessControlIrq.cs
@@ -1,4 +1,4 @@
-namespace LibHac
+namespace LibHac.Npdm
 {
     struct KernelAccessControlIrq
     {
