+++ LibHac/Npdm/SvcName.cs
@@ -1,4 +1,4 @@
-namespace LibHac
+namespace LibHac.Npdm
 {
     enum SvcName
     {
